[Refactor] 유저 정보 TextInput 파일 분리
