+++ components/MyPage/index.js
@@ -1,1 +1,2 @@
 export * from "./LogoutBtn";
+export * from "./UserInfo";
